LPS-53802 Rename
